Use GOOS=js and GOARCH=ecmascript to build user code by default.
GopherJS will use this target going forward. The values where chosen by
the following logic:

  - GOOS describes the targeted API environment, which is generally the
    same as when building wasm: both execute under a browser or nodejs.
    Build tag `js` is shared by GopherJS and wasm and can be used to
    target the general environment.
  - GOARCH describes the low-level emitted code, which is currently
    EcmaScript 5. Using `ecmascript` here differentiates from `wasm`,
    while not referencing GopherJS explicitly. It also leaves us a
    different paths for future upgrades to ES6+ by either changing
    GOARCH or keeping it and making this change implicit.

At the same time, we use fixed js/wasm target when building standard
library packages, which resolves a lot of concerns discussed in
https://github.com/gopherjs/gopherjs/issues/693. Using js/ecmascript is
not feasible because it'll require us to maintain a lot more overlays
and/or rewrite build tags, which leads to the exactly same outcome.

In addition, we always set `gopherjs` build tag to make it easy to
guard sources using GopherJS-specific features. This is similar to `gc`
and `gccgo` build tags respective compilers set.

This change is potentially breaking for users that may have relied on
GOOS=linux in their sources. To make transition easier for them, we
support using GOOS/GOARCH from environment. However, they will be only
applied to the user code. Standard library will still be built with
js/wasm.