Disable CGo when loading packages.
GopherJS doesn't support CGo. This was originally enabled to address
https://github.com/gopherjs/gopherjs/issues/215 by providing a better
error when attempting to build a CGo package (instead of a compiler
panic). However, this approach makes GopherJS refuse to build packages,
which have a CGo and purego versions guarded by the `cgo` build tag:
GopherJS woudl try to load the cgo version and error out.

After this change GopherJS won't attempt to load Go sources with `import
"C"`, and it will load sources with `//go:build !cgo` instead. For
packages that require CGo one of the following errors will be returned:

  - "No buildable sources" if all Go sources import C.
  - "Undefined symbol" for symbols defined in the Go sources that import
    C and that are referenced by other Go source files.
  - "Unknown import path C" if you try to `gopherjs run` a source with
    import C directly.